Update client.ts, clear types
