+++ src/lib/client.ts
@@ -1,131 +1,24 @@
 import { ApiPromise, WsProvider, } from '@polkadot/api'
 import { Vec } from '@polkadot/types/codec'
-import { PeerInfo } from '@polkadot/types/interfaces/system'
 import * as event from '@tauri-apps/api/event'
 import { reactive } from 'vue'
 import { LocalStorage } from 'quasar'
 import { AccountId32 } from '@polkadot/types/interfaces';
-
-// import { Header } from '@polkadot/types/interfaces/runtime'
 import * as process from 'process'
-import mitt, { Emitter } from 'mitt'
-import { FarmedBlock } from './types'
+import { ClientIdentity, emptyClientData, FarmedBlock, NetStatus } from './types'
 import { SubPreDigest } from './customTypes/types'
 import { invoke } from '@tauri-apps/api/tauri'
 
 const tauri = { event, invoke }
 const SUNIT = 1000000000000000000;
-
-export interface NetStatus { peers: Vec<PeerInfo> }
-
-export interface PeerData {
-  status: 'disconnected' | 'unstable' | 'connected' | string
-  name: string // do peers have some string identifier?
-  ip: string
-  receivedBytes: number
-  sentBytes: number
-}
-
-export interface ClientNetwork {
-  status: 'disconnected' | 'unstable' | 'connected' | string
-  peers: PeerData[]
-  details: {
-    // physical network interface
-    // more granular connection information here
-  }
-}
-
-export interface ClientPlot {
-  status: 'active' | 'verifying' | 'corrupted' | 'syncing' | string
-  plotSizeBytes: number // size of the plot file in Bytes
-  plotFile: string // drive directory where the plot file is located
-  details: {
-    // additional information could be placed here
-  }
-}
-
-export interface Block {
-  id: string
-  time: Date
-  transactions: string[]
-  reward: number
-  fees: number
-}
-
-export interface ClientFarming {
-  status: 'active' | 'paused' | string
-  farmed: FarmedBlock[],
-  events: Emitter<any>
-}
-
-export interface ClientData {
-  plot: ClientPlot
-  network: ClientNetwork
-  farming: ClientFarming
-}
-
-export interface ClientIdentity {
-  publicKey: string,
-  mnemonic: string
-}
-
-export const emptyData: ClientData = {
-  plot: { details: {}, plotFile: '', plotSizeBytes: 0, status: '' },
-  farming: { farmed: [], status: '', events: mitt() },
-  network: { details: {}, peers: [], status: '' }
-}
-
-export interface ClientType {
-  api: ApiPromise | null
-  data: ClientData
-  getStatus: {
-    farming: () => void,
-    plot: () => void,
-    network: () => void
-  },
-  do?: { [index: string]: any }
-}
-
-function getStoredBlocks(): FarmedBlock[] {
-  const mined: FarmedBlock[] = []
-  try {
-    const blocks = LocalStorage.getItem('farmedBlocks')
-    if (!blocks) return []
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    for (const [num, block] of Object.entries(blocks)) {
-      mined.push(block as FarmedBlock)
-    }
-  } catch (error) {
-    console.error(error, 'error reading stored blocks')
-  }
-  return mined
-}
-
-function storeBlocks(blocks: FarmedBlock[]): void {
-  const farmed: { [index: string]: FarmedBlock } = {}
-  for (const block of blocks) {
-    farmed[block.id] = block
-  }
-  LocalStorage.set('farmedBlocks', farmed)
-}
-
-function clearStored(): void {
-  try {
-    LocalStorage.remove('farmedBlocks')
-  } catch (error) {
-    console.error('error clearing mined blocks')
-  }
-}
-
-//TODO should be refactored to not rely on the .init() method to be valid
-export class Client {  
-  protected clientIsReady = false;
+export class Client {
   protected wsProvider = new WsProvider("ws://localhost:9944");
   protected api: ApiPromise = new ApiPromise({ provider: this.wsProvider });
-  protected farmed = getStoredBlocks();
+  protected farmed: FarmedBlock[] = [];
   protected clearTauriDestroy: event.UnlistenFn = () => { };
   protected unsubscribe: event.UnlistenFn = () => { };
-  data = reactive(emptyData);
+  data = reactive(emptyClientData);
+  protected clientStarted = false;
 
   status = {
     farming: (): void => { }, // TODO return some farming status info
@@ -137,7 +30,6 @@ export class Client {
   }
   do = {
     blockSubscription: {
-      clearStored,
       stopOnReload(): void {
         this.stop()
       },
@@ -173,14 +65,14 @@ export class Client {
               this.data.farming.farmed = [block].concat(
                 this.data.farming.farmed
               );
-              storeBlocks(this.farmed);
+              this.storeBlocks(this.data.farming.farmed)
             }
           })
         process.on('beforeExit', this.do.blockSubscription.stopOnReload)
         window.addEventListener('unload', this.do.blockSubscription.stopOnReload)
         this.clearTauriDestroy = await tauri.event.once('tauri://destroyed', () => {
           console.log('Destroyed event!')
-          storeBlocks(this.data.farming.farmed)
+          this.storeBlocks(this.data.farming.farmed)
         })
       },
       stop: (): void => {
@@ -188,37 +80,83 @@ export class Client {
         this.unsubscribe()
         try {
           this.clearTauriDestroy()
-          storeBlocks(this.data.farming.farmed)
+          this.storeBlocks(this.data.farming.farmed)
           window.removeEventListener('unload', this.do.blockSubscription.stopOnReload)
         } catch (error) {
           console.error(error)
         }
       },
-      runTest(): void {
+      startSubcriptions(): void {
         this.start()
       }
     }
   }
   constructor() {
-    this.data.farming.farmed = this.farmed
   }
-  isConnected(): boolean{
-    return this.clientIsReady
-  }
-  async init(): Promise<void> {
-    this.clientIsReady = true
-    this.api = await ApiPromise.create({
-      provider: this.wsProvider, types: {
-        Solution: {
-          public_key: 'AccountId32'
-        },
-        SubPreDigest: {
-          slot: 'u64',
-          solution: 'Solution'
+
+  public async init(clear: boolean = false, farmerPublicKey?: string | undefined , mnemonic?: string | undefined ): Promise<void> {
+    if (farmerPublicKey && mnemonic) {
+      LocalStorage.set("farmerPublicKey", farmerPublicKey)
+      LocalStorage.set("mnemonic", mnemonic)
+    }
+    
+    if (clear)
+      this.clearStoredBlocks()
+    else
+      this.loadStoredBlocks()
+
+    if (!this.clientStarted) {
+      this.api = await ApiPromise.create({
+        provider: this.wsProvider, types: {
+          Solution: {
+            public_key: 'AccountId32'
+          },
+          SubPreDigest: {
+            slot: 'u64',
+            solution: 'Solution'
+          }
         }
+      })
+      this.do.blockSubscription.startSubcriptions()
+    }
+
+    this.clientStarted = true;
+  }
+
+  private getStoredBlocks(): FarmedBlock[] {
+    const mined: FarmedBlock[] = []
+    try {
+      const blocks = LocalStorage.getItem('farmedBlocks')
+      if (!blocks) return []
+      // eslint-disable-next-line @typescript-eslint/no-unused-vars
+      for (const [num, block] of Object.entries(blocks)) {
+        mined.push(block as FarmedBlock)
       }
-    })
-    this.do.blockSubscription.runTest()
+    } catch (error) {
+      console.error(error, 'error reading stored blocks')
+    }
+    return mined
+  }
+
+  private storeBlocks(blocks: FarmedBlock[]): void {
+    const farmed: { [index: string]: FarmedBlock } = {}
+    for (const block of blocks) {
+      farmed[block.id] = block
+    }
+    LocalStorage.set('farmedBlocks', farmed)
+  }
+
+  private clearStoredBlocks(): void {
+    try {
+      LocalStorage.remove('farmedBlocks')
+    } catch (error) {
+      console.error('error clearing mined blocks')
+    }
+  }
+
+  private loadStoredBlocks(): void {
+    this.farmed = this.getStoredBlocks();
+    this.data.farming.farmed = this.farmed
   }
 }
